Moved html page generation functions from ciftify/utilities.py to ciftify/html.py
