+++ ciftify/utilities.py
@@ -366,98 +366,3 @@ def run(cmd, dryrun=False, echo=True, supress_stdout = False):
         logger.warning(err)
 
     return p.returncode
-
-def add_page_header(html_page, qc_config, page_subject, subject=None, active_link=None,
-        path='', title=None):
-    """
-    Adds a QC page header for a 'snap' or index page in the cifti_vis_* scripts.
-    """
-
-    if title is None:
-        title = page_subject
-
-    first_line = '<!DOCTYPE html>\n<HTML><TITLE>'
-    if subject is not None:
-        first_line = first_line + " QC {}".format(subject)
-    first_line = first_line + " {} </TITLE>\n".format(title)
-
-    html_page.write(first_line)
-    ciftify.html.write_header(html_page)
-
-    html_page.write('<body>\n')
-    nav_list = get_navigation_list(qc_config, path)
-    ciftify.html.write_navbar(html_page, page_subject, nav_list, active_link)
-
-    if subject is not None:
-        html_page.write('\n<h1>QC {} {}</h1>\n'.format(subject, page_subject))
-
-def get_navigation_list(qc_config, path=''):
-    nav_list = [{'href': '', 'label':'View:'}]
-
-    for image in qc_config.images:
-        if image.make_index:
-            image_path = os.path.join(path, '{}.html'.format(image.name))
-            nav_list.append({ 'href': image_path,
-                              'label': image.name})
-
-    index_path = os.path.join(path, 'index.html')
-    nav_list.append({'href': index_path, 'label':'Index'})
-
-    return nav_list
-
-def add_images(qc_page, qc_dir, image_list, scene_file):
-    """
-    Takes a list of scenes and montages, generates them, and adds them to
-    qc_page.
-    """
-    for image in image_list:
-        pic_name = "{}.png".format(image.name)
-        ciftify.html.add_image(qc_page, 12, pic_name, pic_name, "")
-        output_path = os.path.join(qc_dir, pic_name)
-        image.make_image(output_path, scene_file)
-
-def add_image_and_subject_index(index_page, config, subjects, page_subject):
-    index_page.write('<h1>{} Index</h1>\n'.format(page_subject))
-    index_page.write('<h2>All subjects together</h2>\n')
-    index_page.write('<ul>\n  ')
-    for image in config.images:
-        if not image.make_index:
-            continue
-        index_page.write('<li><a href="{}.html">{} View</a>' \
-                '</li>\n'.format(image.name, image.name))
-    index_page.write('</ul>\n')
-    index_page.write('<h2>Subject Pages</h2>\n')
-    index_page.write('<ul>\n  ')
-    for subject in subjects:
-        index_page.write('<li><a href="{}/qc.html">{}</a>' \
-                '</li>\n'.format(subject, subject))
-    index_page.write('</ul>\n')
-    index_page.write('</body>')
-
-def write_image_index(qc_dir, subjects, qc_config, page_subject, image_name,
-        title=None, colwidth=12):
-    '''
-    Writes html file with all subjects for one pic shown together
-    '''
-    pic_name = "{}.png".format(image_name)
-    html_name = "{}.html".format(image_name)
-    # open the file
-    html_index = os.path.join(qc_dir, html_name)
-    with open(html_index, 'w') as image_page:
-        add_page_header(image_page, qc_config, page_subject, title=title,
-                active_link=html_name)
-        ## add the main title
-        if title is not None:
-            image_page.write('<h1>{}</h1>\n'.format(title))
-        for subject in subjects:
-            add_image_and_subject_page_link(image_page, subject, pic_name,
-                    colwidth)
-        ## close the html page
-        image_page.write('</body>\n')
-
-def add_image_and_subject_page_link(image_page, subject, pic_name, colwidth):
-    image_page.write('<div class="container" style="width: 100%;">')
-    subject_page = os.path.join('{}'.format(subject), 'qc.html')
-    pic = os.path.join('{}'.format(subject), pic_name)
-    ciftify.html.add_image(image_page, colwidth, subject_page, pic, subject)
-    image_page.write('</div>\n</br>')
